Added Games DB Curl Scripts
Added curl scripts for:
games-create, which creates a game. It requires an auth TOKEN, and a TITLE as a string. Run script with sh scripts/games-library/games-create.sh
games-destroy, which deletes a game from the database. It requires the game ID. Run script with sh scripts/games-library/games-destroy.sh
games-index, which returns the full list of games within the user's database. Run script with #sh scripts/games-library/games-index.sh
games-show, which shows a single game based on the game ID. It requires the game ID. Run script with #sh scripts/games-library/games-show.sh
games-update, which updates a users game in the database. It requires the game ID and the updated title as a string. Run script with #sh scripts/games-library/games-update.sh